requirements.txt
langchain-google-genai
langchain-openai
requests
python-dotenv
typing-extensions
numpy
scipy
twilio
faiss-cpu
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
idna==3.10
multidict==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.38.6
twilio==9.3.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
websockets==13.1
yarl==1.12.1
langchain
langchain_community
beautifulsoup4
motor
python-multipart